CA-147944: Use GB for VM Memory if large values - Memory Tab uses GB units for values greater or equal to 1 GB, otherwise show in MB; - Search Tab displays values in GB or MB in the same as the Memory Tab (e.g. 512 MB of 1 GB, 256 MB of 512 MB, 2.5 GB of 16 GB). - The shiny bar present in the Memory Tab and in Memory Settings dialog shows the scaling in the following way: If smaller than 1 GB, then show as before, else show only labels with values multiples of half a GB. - The units used in Memory Setting Dialog are set depending on the static_max. If it is greater or equal to 1 GB, then the units are GB, else MB. The user does not have the possibility of changing them.

+++ XenAdmin/Controls/CustomDataGraph/DataSet.cs
@@ -206,7 +206,7 @@ namespace XenAdmin.Controls.CustomDataGraph
                     dataSet.Name = Helpers.GetFriendlyDataSourceName("memory_internal_used", dataSet.XenObject);
                     VM vm = (VM)xo;
                     VM_metrics metrics = vm.Connection.Resolve(vm.metrics);
-                    long max = metrics != null ? metrics.memory_actual : vm.memory_dynamic_max;
+                    long max = metrics != null ? metrics.memory_actual : (long)vm.memory_dynamic_max;
                     dataSet.CustomYRange = new DataRange(max, 0, max / 10d, Unit.Bytes, RangeScaleMode.Delegate)
                     {
                         UpdateMax = dataSet.GetMemoryMax,
@@ -396,7 +396,7 @@ namespace XenAdmin.Controls.CustomDataGraph
             {
                 VM vm = (VM)xmo;
                 VM_metrics metrics = vm.Connection.Resolve(vm.metrics);
-                return ((metrics != null ? metrics.memory_actual : vm.memory_dynamic_max) / 10d);
+                return ((metrics != null ? metrics.memory_actual : (long)vm.memory_dynamic_max) / 10d);
             }
             return 10;
         }
